fix(runtime): project runtime tree through shared status
