+++ crates/harness-server/src/http/misc_routes.rs
@@ -6,31 +6,18 @@ use axum::{
     Json,
 };
 use harness_protocol::methods::RpcRequest;
-use serde_json::{json, Value};
-use std::collections::{BTreeMap, BTreeSet};
+use serde_json::json;
 use std::path::{Path as StdPath, PathBuf};
 use std::sync::Arc;
 
 use super::{state::AppState, task_routes};
 use crate::runtime_projection::RuntimeWorkflowProjection;
 use crate::{router, task_runner};
-use harness_workflow::runtime::store::{RuntimeEventSummary, RuntimeJobCompactRecord};
-use harness_workflow::runtime::{
-    runtime_job_running_lease_state_at, RuntimeEvent, RuntimeJob, RuntimeJobStatus,
-    WorkflowCommand, WorkflowCommandRecord, WorkflowDecisionRecord, WorkflowEvent,
-    WorkflowInstance, WorkflowLease,
-};
+use harness_workflow::runtime::WorkflowInstance;
 
-const ACTIVITY_RESULT_ENVELOPE_ARTIFACT_TYPE: &str = "activity_result_envelope";
-const ACTIVITY_RESULT_ENVELOPE_SCHEMA: &str = "harness.runtime.activity_result_envelope.v1";
-const WORKFLOW_RUNTIME_TREE_DEFAULT_LIMIT: i64 = 100;
-const WORKFLOW_RUNTIME_TREE_MAX_LIMIT: i64 = 500;
-const WORKFLOW_RUNTIME_TREE_DEFAULT_JOB_LIMIT: i64 = 5;
-const WORKFLOW_RUNTIME_TREE_MAX_JOB_LIMIT: i64 = 50;
-const WORKFLOW_RUNTIME_TREE_DEFAULT_COMMAND_LIMIT: i64 = 5;
-const WORKFLOW_RUNTIME_TREE_MAX_COMMAND_LIMIT: i64 = 50;
-const WORKFLOW_RUNTIME_TREE_DEFAULT_REJECTED_DECISION_LIMIT: i64 = 3;
-const WORKFLOW_RUNTIME_TREE_MAX_REJECTED_DECISION_LIMIT: i64 = 20;
+#[path = "misc_routes_runtime_tree.rs"]
+mod runtime_tree;
+pub(crate) use runtime_tree::get_workflow_runtime_tree;
 
 #[derive(Debug)]
 struct IntakeRecentDispatch {
@@ -146,288 +133,6 @@ pub(crate) struct ProjectWorkflowByProjectQuery {
     pub repo: Option<String>,
 }
 
-#[derive(Debug, serde::Deserialize)]
-pub(crate) struct WorkflowRuntimeTreeQuery {
-    pub project_id: Option<String>,
-    pub limit: Option<i64>,
-    pub offset: Option<i64>,
-    pub job_limit: Option<i64>,
-    pub command_limit: Option<i64>,
-    pub rejected_decision_limit: Option<i64>,
-    pub summary_only: Option<bool>,
-    pub detail: Option<WorkflowRuntimeTreeDetail>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, serde::Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub(crate) enum WorkflowRuntimeTreeDetail {
-    Compact,
-    Full,
-}
-
-impl WorkflowRuntimeTreeDetail {
-    fn as_str(self) -> &'static str {
-        match self {
-            Self::Compact => "compact",
-            Self::Full => "full",
-        }
-    }
-}
-
-#[derive(Debug, Clone, serde::Serialize)]
-struct WorkflowRuntimeCommandNode {
-    pub id: String,
-    pub workflow_id: String,
-    pub decision_id: Option<String>,
-    pub status: String,
-    pub command: WorkflowCommand,
-    pub runtime_job_count: usize,
-    pub runtime_jobs: Vec<WorkflowRuntimeJobNode>,
-    pub created_at: chrono::DateTime<chrono::Utc>,
-    pub updated_at: chrono::DateTime<chrono::Utc>,
-}
-
-impl WorkflowRuntimeCommandNode {
-    fn new(
-        record: WorkflowCommandRecord,
-        runtime_job_count: usize,
-        runtime_jobs: Vec<WorkflowRuntimeJobNode>,
-    ) -> Self {
-        Self {
-            id: record.id,
-            workflow_id: record.workflow_id,
-            decision_id: record.decision_id,
-            status: record.status.to_string(),
-            command: record.command,
-            runtime_job_count,
-            runtime_jobs,
-            created_at: record.created_at,
-            updated_at: record.updated_at,
-        }
-    }
-}
-
-#[derive(Debug, Clone, serde::Serialize)]
-struct WorkflowRuntimeJobNode {
-    pub id: String,
-    pub command_id: String,
-    pub runtime_kind: harness_workflow::runtime::RuntimeKind,
-    pub runtime_profile: String,
-    pub status: RuntimeJobStatus,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub lease: Option<WorkflowLease>,
-    pub lease_generation: u64,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub input: Option<Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub output: Option<Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub error: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub not_before: Option<chrono::DateTime<chrono::Utc>>,
-    pub created_at: chrono::DateTime<chrono::Utc>,
-    pub updated_at: chrono::DateTime<chrono::Utc>,
-    pub runtime_event_count: usize,
-    pub latest_runtime_event_type: Option<String>,
-    pub prompt_packet_digest: Option<String>,
-    pub activity_result_envelope: Option<Value>,
-    pub lease_state: Option<&'static str>,
-    pub in_flight_model_turn: bool,
-    pub last_runtime_observation_at: Option<chrono::DateTime<chrono::Utc>>,
-}
-
-impl WorkflowRuntimeJobNode {
-    fn full(job: RuntimeJob, runtime_events: Vec<RuntimeEvent>) -> Self {
-        let latest_runtime_event_type = runtime_events.last().map(|event| event.event_type.clone());
-        let prompt_packet_digest = prompt_packet_digest_from_events(&runtime_events);
-        let activity_result_envelope = activity_result_envelope_from_job(&job);
-        let lease_state = runtime_job_running_lease_state_at(&job, chrono::Utc::now())
-            .map(|state| state.status_label());
-        let in_flight_model_turn = runtime_job_has_in_flight_model_turn(&job, &runtime_events);
-        let last_runtime_observation_at = last_runtime_observation_at(&job, &runtime_events);
-        Self {
-            id: job.id,
-            command_id: job.command_id,
-            runtime_kind: job.runtime_kind,
-            runtime_profile: job.runtime_profile,
-            status: job.status,
-            lease: job.lease,
-            lease_generation: job.lease_generation,
-            input: Some(job.input),
-            output: job.output,
-            error: job.error,
-            not_before: job.not_before,
-            created_at: job.created_at,
-            updated_at: job.updated_at,
-            runtime_event_count: runtime_events.len(),
-            latest_runtime_event_type,
-            prompt_packet_digest,
-            activity_result_envelope,
-            lease_state,
-            in_flight_model_turn,
-            last_runtime_observation_at,
-        }
-    }
-
-    fn compact(job: RuntimeJobCompactRecord, runtime_summary: RuntimeEventSummary) -> Self {
-        let now = chrono::Utc::now();
-        let lease_state = if job.status == RuntimeJobStatus::Running {
-            Some(match job.lease.as_ref() {
-                Some(lease) if lease.expires_at > now => "active_leased",
-                Some(_) => "expired_lease",
-                None => "missing_lease",
-            })
-        } else {
-            None
-        };
-        let latest_turn_sequence = runtime_summary.latest_turn_sequence.unwrap_or(0);
-        let latest_activity_result_sequence =
-            runtime_summary.latest_activity_result_sequence.unwrap_or(0);
-        let in_flight_model_turn = job.status == RuntimeJobStatus::Running
-            && latest_turn_sequence > latest_activity_result_sequence;
-        let last_runtime_observation_at = if job.status == RuntimeJobStatus::Running {
-            Some(
-                runtime_summary
-                    .latest_runtime_event_at
-                    .map(|created_at| created_at.max(job.updated_at))
-                    .unwrap_or(job.updated_at),
-            )
-        } else {
-            runtime_summary.latest_runtime_event_at
-        };
-        Self {
-            id: job.id,
-            command_id: job.command_id,
-            runtime_kind: job.runtime_kind,
-            runtime_profile: job.runtime_profile,
-            status: job.status,
-            lease: job.lease,
-            lease_generation: job.lease_generation,
-            input: None,
-            output: None,
-            error: job.error,
-            not_before: job.not_before,
-            created_at: job.created_at,
-            updated_at: job.updated_at,
-            runtime_event_count: runtime_summary.runtime_event_count,
-            latest_runtime_event_type: runtime_summary.latest_runtime_event_type,
-            prompt_packet_digest: runtime_summary.prompt_packet_digest,
-            activity_result_envelope: None,
-            lease_state,
-            in_flight_model_turn,
-            last_runtime_observation_at,
-        }
-    }
-}
-
-#[derive(Debug, Clone, serde::Serialize)]
-struct WorkflowRuntimeTreeNode {
-    pub workflow: WorkflowInstance,
-    pub runtime_job_count: usize,
-    pub event_count: usize,
-    pub decision_count: usize,
-    pub rejected_decision_count: usize,
-    pub command_count: usize,
-    pub events: Vec<WorkflowEvent>,
-    pub decisions: Vec<WorkflowDecisionRecord>,
-    pub commands: Vec<WorkflowRuntimeCommandNode>,
-    pub children: Vec<WorkflowRuntimeTreeNode>,
-}
-
-#[derive(Debug, serde::Serialize)]
-struct WorkflowRuntimeTreeResponse {
-    pub workflows: Vec<WorkflowRuntimeTreeNode>,
-    pub total_workflows: usize,
-    pub pagination: WorkflowRuntimeTreePagination,
-    pub summary: WorkflowRuntimeTreeSummary,
-}
-
-#[derive(Debug, serde::Serialize)]
-struct WorkflowRuntimeTreePagination {
-    pub limit: usize,
-    pub offset: usize,
-    pub returned: usize,
-    pub total: usize,
-    pub has_more: bool,
-    pub next_offset: Option<usize>,
-    pub job_limit: usize,
-    pub command_limit: Option<usize>,
-    pub detail: &'static str,
-    pub summary_only: bool,
-}
-
-impl WorkflowRuntimeTreePagination {
-    fn new(
-        limit: i64,
-        offset: i64,
-        returned: usize,
-        total: i64,
-        job_limit: usize,
-        command_limit: Option<usize>,
-        detail: WorkflowRuntimeTreeDetail,
-        summary_only: bool,
-    ) -> Self {
-        let limit = limit.max(1) as usize;
-        let offset = offset.max(0) as usize;
-        let total = total.max(0) as usize;
-        let next_offset = offset.saturating_add(returned);
-        let has_more = !summary_only && next_offset < total;
-        Self {
-            limit,
-            offset,
-            returned,
-            total,
-            has_more,
-            next_offset: has_more.then_some(next_offset),
-            job_limit,
-            command_limit,
-            detail: detail.as_str(),
-            summary_only,
-        }
-    }
-}
-
-#[derive(Debug, Default, serde::Serialize)]
-struct WorkflowRuntimeTreeSummary {
-    pub total_commands: usize,
-    pub total_runtime_jobs: usize,
-    pub command_statuses: BTreeMap<String, usize>,
-    pub runtime_job_statuses: BTreeMap<String, usize>,
-    pub running_job_lease_statuses: BTreeMap<String, usize>,
-    pub activity_outcomes: BTreeMap<String, usize>,
-    pub jobs_without_activity_envelope: usize,
-}
-
-#[derive(Debug, Clone, Copy)]
-enum WorkflowRuntimeTreeMode {
-    SummaryOnly,
-    Compact {
-        command_limit: usize,
-        rejected_decision_limit: usize,
-    },
-    Full,
-}
-
-impl WorkflowRuntimeTreeMode {
-    fn detail(self) -> WorkflowRuntimeTreeDetail {
-        match self {
-            Self::Full => WorkflowRuntimeTreeDetail::Full,
-            Self::SummaryOnly | Self::Compact { .. } => WorkflowRuntimeTreeDetail::Compact,
-        }
-    }
-
-    fn summary_only(self) -> bool {
-        matches!(self, Self::SummaryOnly)
-    }
-
-    fn command_limit(self) -> Option<usize> {
-        match self {
-            Self::Compact { command_limit, .. } => Some(command_limit),
-            Self::SummaryOnly | Self::Full => None,
-        }
-    }
-}
-
 pub(crate) async fn get_issue_workflow_by_issue(
     State(state): State<Arc<AppState>>,
     Query(query): Query<IssueWorkflowByIssueQuery>,
@@ -514,520 +219,6 @@ pub(crate) async fn get_project_workflow_by_project(
             .into_response(),
     }
 }
-
-pub(crate) async fn get_workflow_runtime_tree(
-    State(state): State<Arc<AppState>>,
-    Query(query): Query<WorkflowRuntimeTreeQuery>,
-) -> Response {
-    let Some(store) = state.core.workflow_runtime_store.as_ref() else {
-        return (
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(json!({ "error": "workflow runtime store unavailable" })),
-        )
-            .into_response();
-    };
-    let limit = query
-        .limit
-        .unwrap_or(WORKFLOW_RUNTIME_TREE_DEFAULT_LIMIT)
-        .clamp(1, WORKFLOW_RUNTIME_TREE_MAX_LIMIT);
-    let offset = query.offset.unwrap_or(0).max(0);
-    let job_limit = query
-        .job_limit
-        .unwrap_or(WORKFLOW_RUNTIME_TREE_DEFAULT_JOB_LIMIT)
-        .clamp(0, WORKFLOW_RUNTIME_TREE_MAX_JOB_LIMIT);
-    let command_limit = query
-        .command_limit
-        .unwrap_or(WORKFLOW_RUNTIME_TREE_DEFAULT_COMMAND_LIMIT)
-        .clamp(0, WORKFLOW_RUNTIME_TREE_MAX_COMMAND_LIMIT);
-    let rejected_decision_limit = query
-        .rejected_decision_limit
-        .unwrap_or(WORKFLOW_RUNTIME_TREE_DEFAULT_REJECTED_DECISION_LIMIT)
-        .clamp(0, WORKFLOW_RUNTIME_TREE_MAX_REJECTED_DECISION_LIMIT);
-    let detail = query.detail.unwrap_or(WorkflowRuntimeTreeDetail::Compact);
-    let mode = if query.summary_only.unwrap_or(false) {
-        WorkflowRuntimeTreeMode::SummaryOnly
-    } else {
-        match detail {
-            WorkflowRuntimeTreeDetail::Compact => WorkflowRuntimeTreeMode::Compact {
-                command_limit: command_limit as usize,
-                rejected_decision_limit: rejected_decision_limit as usize,
-            },
-            WorkflowRuntimeTreeDetail::Full => WorkflowRuntimeTreeMode::Full,
-        }
-    };
-    match store
-        .list_instances_page(query.project_id.as_deref(), limit, offset)
-        .await
-    {
-        Ok(page) => match build_workflow_runtime_tree(
-            store,
-            page,
-            query.project_id.as_deref(),
-            job_limit as usize,
-            mode,
-        )
-        .await
-        {
-            Ok(response) => (StatusCode::OK, Json(response)).into_response(),
-            Err(e) => (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(json!({ "error": e.to_string() })),
-            )
-                .into_response(),
-        },
-        Err(e) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(json!({ "error": e.to_string() })),
-        )
-            .into_response(),
-    }
-}
-
-async fn build_workflow_runtime_tree(
-    store: &harness_workflow::runtime::WorkflowRuntimeStore,
-    page: harness_workflow::runtime::store::WorkflowInstancePage,
-    project_id: Option<&str>,
-    job_limit: usize,
-    mode: WorkflowRuntimeTreeMode,
-) -> anyhow::Result<WorkflowRuntimeTreeResponse> {
-    let instances = page.instances;
-    let workflow_ids: Vec<String> = instances
-        .iter()
-        .map(|instance| instance.id.clone())
-        .collect();
-    let summary = workflow_runtime_tree_summary(store, project_id).await?;
-    if mode.summary_only() {
-        return Ok(WorkflowRuntimeTreeResponse {
-            total_workflows: page.total.max(0) as usize,
-            pagination: WorkflowRuntimeTreePagination::new(
-                page.limit,
-                page.offset,
-                0,
-                page.total,
-                job_limit,
-                mode.command_limit(),
-                mode.detail(),
-                true,
-            ),
-            summary,
-            workflows: Vec::new(),
-        });
-    }
-
-    let (by_id, children_by_parent) = match mode {
-        WorkflowRuntimeTreeMode::Full => {
-            build_full_workflow_runtime_nodes(store, instances, &workflow_ids, job_limit).await?
-        }
-        WorkflowRuntimeTreeMode::Compact {
-            command_limit,
-            rejected_decision_limit,
-        } => {
-            build_compact_workflow_runtime_nodes(
-                store,
-                instances,
-                &workflow_ids,
-                job_limit,
-                command_limit,
-                rejected_decision_limit,
-            )
-            .await?
-        }
-        WorkflowRuntimeTreeMode::SummaryOnly => unreachable!("summary-only returned above"),
-    };
-
-    let mut root_ids = Vec::new();
-    for (workflow_id, node) in &by_id {
-        let parent_present = node
-            .workflow
-            .parent_workflow_id
-            .as_ref()
-            .is_some_and(|parent_id| by_id.contains_key(parent_id));
-        if !parent_present {
-            root_ids.push(workflow_id.clone());
-        }
-    }
-    if root_ids.is_empty() {
-        root_ids.extend(by_id.keys().cloned());
-    }
-
-    let mut workflows = Vec::new();
-    for root_id in root_ids {
-        let mut path = BTreeSet::new();
-        if let Some(node) =
-            attach_workflow_children(&root_id, &by_id, &children_by_parent, &mut path)
-        {
-            workflows.push(node);
-        }
-    }
-
-    Ok(WorkflowRuntimeTreeResponse {
-        total_workflows: page.total.max(0) as usize,
-        pagination: WorkflowRuntimeTreePagination::new(
-            page.limit,
-            page.offset,
-            by_id.len(),
-            page.total,
-            job_limit,
-            mode.command_limit(),
-            mode.detail(),
-            false,
-        ),
-        summary,
-        workflows,
-    })
-}
-
-async fn workflow_runtime_tree_summary(
-    store: &harness_workflow::runtime::WorkflowRuntimeStore,
-    project_id: Option<&str>,
-) -> anyhow::Result<WorkflowRuntimeTreeSummary> {
-    let aggregate_summary = store
-        .runtime_summary_counts_for_instances(
-            project_id,
-            ACTIVITY_RESULT_ENVELOPE_ARTIFACT_TYPE,
-            ACTIVITY_RESULT_ENVELOPE_SCHEMA,
-        )
-        .await?;
-    Ok(WorkflowRuntimeTreeSummary {
-        total_commands: aggregate_summary.total_commands,
-        total_runtime_jobs: aggregate_summary.total_runtime_jobs,
-        command_statuses: aggregate_summary.command_statuses,
-        runtime_job_statuses: aggregate_summary.runtime_job_statuses,
-        running_job_lease_statuses: aggregate_summary.running_job_lease_statuses,
-        activity_outcomes: aggregate_summary.activity_outcomes,
-        jobs_without_activity_envelope: aggregate_summary.jobs_without_activity_envelope,
-    })
-}
-
-async fn build_full_workflow_runtime_nodes(
-    store: &harness_workflow::runtime::WorkflowRuntimeStore,
-    instances: Vec<WorkflowInstance>,
-    workflow_ids: &[String],
-    job_limit: usize,
-) -> anyhow::Result<(
-    BTreeMap<String, WorkflowRuntimeTreeNode>,
-    BTreeMap<String, Vec<String>>,
-)> {
-    let mut events_by_workflow = store.events_for_workflows(workflow_ids).await?;
-    let mut decisions_by_workflow = store.decisions_for_workflows(workflow_ids).await?;
-    let mut commands_by_workflow = store.commands_for_workflows(workflow_ids).await?;
-    let command_ids: Vec<String> = commands_by_workflow
-        .values()
-        .flat_map(|commands| commands.iter().map(|command| command.id.clone()))
-        .collect();
-    let runtime_job_counts_by_command = store.runtime_job_counts_for_commands(&command_ids).await?;
-    let mut runtime_jobs_by_command = store
-        .runtime_jobs_for_commands_limited(&command_ids, job_limit as i64)
-        .await?;
-    let runtime_job_ids: Vec<String> = runtime_jobs_by_command
-        .values()
-        .flat_map(|jobs| jobs.iter().map(|job| job.id.clone()))
-        .collect();
-    let mut runtime_events_by_job = store.runtime_events_for_jobs(&runtime_job_ids).await?;
-
-    let mut by_id = BTreeMap::new();
-    let mut children_by_parent: BTreeMap<String, Vec<String>> = BTreeMap::new();
-    for instance in instances {
-        let workflow_id = instance.id.clone();
-        if let Some(parent_id) = instance.parent_workflow_id.as_ref() {
-            children_by_parent
-                .entry(parent_id.clone())
-                .or_default()
-                .push(workflow_id.clone());
-        }
-        let events = events_by_workflow.remove(&workflow_id).unwrap_or_default();
-        let decisions = decisions_by_workflow
-            .remove(&workflow_id)
-            .unwrap_or_default();
-        let rejected_decision_count = decisions
-            .iter()
-            .filter(|decision| !decision.accepted)
-            .count();
-        let command_records = commands_by_workflow
-            .remove(&workflow_id)
-            .unwrap_or_default();
-        let command_count = command_records.len();
-        let commands: Vec<WorkflowRuntimeCommandNode> = command_records
-            .into_iter()
-            .map(|command| {
-                let runtime_job_count = runtime_job_counts_by_command
-                    .get(&command.id)
-                    .copied()
-                    .unwrap_or_default();
-                let runtime_jobs = runtime_jobs_by_command
-                    .remove(&command.id)
-                    .unwrap_or_default()
-                    .into_iter()
-                    .map(|job| {
-                        let runtime_events =
-                            runtime_events_by_job.remove(&job.id).unwrap_or_default();
-                        WorkflowRuntimeJobNode::full(job, runtime_events)
-                    })
-                    .collect();
-                WorkflowRuntimeCommandNode::new(command, runtime_job_count, runtime_jobs)
-            })
-            .collect();
-        let runtime_job_count = commands
-            .iter()
-            .map(|command| command.runtime_job_count)
-            .sum();
-        by_id.insert(
-            workflow_id,
-            WorkflowRuntimeTreeNode {
-                workflow: instance,
-                runtime_job_count,
-                event_count: events.len(),
-                decision_count: decisions.len(),
-                rejected_decision_count,
-                command_count,
-                events,
-                decisions,
-                commands,
-                children: Vec::new(),
-            },
-        );
-    }
-    Ok((by_id, children_by_parent))
-}
-
-async fn build_compact_workflow_runtime_nodes(
-    store: &harness_workflow::runtime::WorkflowRuntimeStore,
-    instances: Vec<WorkflowInstance>,
-    workflow_ids: &[String],
-    job_limit: usize,
-    command_limit: usize,
-    rejected_decision_limit: usize,
-) -> anyhow::Result<(
-    BTreeMap<String, WorkflowRuntimeTreeNode>,
-    BTreeMap<String, Vec<String>>,
-)> {
-    let detail_counts_by_workflow = store.detail_counts_for_workflows(workflow_ids).await?;
-    let mut decisions_by_workflow = store
-        .rejected_decisions_for_workflows_limited(workflow_ids, rejected_decision_limit as i64)
-        .await?;
-    let mut commands_by_workflow = store
-        .commands_for_workflows_limited(workflow_ids, command_limit as i64)
-        .await?;
-    let command_ids: Vec<String> = commands_by_workflow
-        .values()
-        .flat_map(|commands| commands.iter().map(|command| command.id.clone()))
-        .collect();
-    let runtime_job_counts_by_command = store.runtime_job_counts_for_commands(&command_ids).await?;
-    let mut runtime_jobs_by_command = store
-        .compact_runtime_jobs_for_commands_limited(&command_ids, job_limit as i64)
-        .await?;
-    let runtime_job_ids: Vec<String> = runtime_jobs_by_command
-        .values()
-        .flat_map(|jobs| jobs.iter().map(|job| job.id.clone()))
-        .collect();
-    let mut runtime_event_summaries_by_job = store
-        .runtime_event_summaries_for_jobs(&runtime_job_ids)
-        .await?;
-
-    let mut by_id = BTreeMap::new();
-    let mut children_by_parent: BTreeMap<String, Vec<String>> = BTreeMap::new();
-    for instance in instances {
-        let workflow_id = instance.id.clone();
-        if let Some(parent_id) = instance.parent_workflow_id.as_ref() {
-            children_by_parent
-                .entry(parent_id.clone())
-                .or_default()
-                .push(workflow_id.clone());
-        }
-        let detail_counts = detail_counts_by_workflow
-            .get(&workflow_id)
-            .cloned()
-            .unwrap_or_default();
-        let decisions = decisions_by_workflow
-            .remove(&workflow_id)
-            .unwrap_or_default();
-        let commands: Vec<WorkflowRuntimeCommandNode> = commands_by_workflow
-            .remove(&workflow_id)
-            .unwrap_or_default()
-            .into_iter()
-            .map(|mut command| {
-                command.command = compact_workflow_command(command.command);
-                let runtime_job_count = runtime_job_counts_by_command
-                    .get(&command.id)
-                    .copied()
-                    .unwrap_or_default();
-                let runtime_jobs = runtime_jobs_by_command
-                    .remove(&command.id)
-                    .unwrap_or_default()
-                    .into_iter()
-                    .map(|job| {
-                        let runtime_summary = runtime_event_summaries_by_job
-                            .remove(&job.id)
-                            .unwrap_or_else(|| RuntimeEventSummary {
-                                runtime_job_id: job.id.clone(),
-                                ..Default::default()
-                            });
-                        WorkflowRuntimeJobNode::compact(job, runtime_summary)
-                    })
-                    .collect();
-                WorkflowRuntimeCommandNode::new(command, runtime_job_count, runtime_jobs)
-            })
-            .collect();
-        by_id.insert(
-            workflow_id,
-            WorkflowRuntimeTreeNode {
-                workflow: instance,
-                runtime_job_count: detail_counts.runtime_job_count,
-                event_count: detail_counts.event_count,
-                decision_count: detail_counts.decision_count,
-                rejected_decision_count: detail_counts.rejected_decision_count,
-                command_count: detail_counts.command_count,
-                events: Vec::new(),
-                decisions,
-                commands,
-                children: Vec::new(),
-            },
-        );
-    }
-    Ok((by_id, children_by_parent))
-}
-
-fn compact_workflow_command(command: WorkflowCommand) -> WorkflowCommand {
-    let WorkflowCommand {
-        command_type,
-        dedupe_key,
-        command,
-    } = command;
-    let Value::Object(object) = command else {
-        return WorkflowCommand {
-            command_type,
-            dedupe_key,
-            command,
-        };
-    };
-
-    let mut payload = serde_json::Map::new();
-    for key in [
-        "activity",
-        "definition_id",
-        "subject_key",
-        "pr_number",
-        "pr_url",
-        "reason",
-        "concern",
-    ] {
-        if let Some(value) = object.get(key) {
-            payload.insert(key.to_string(), value.clone());
-        }
-    }
-    WorkflowCommand {
-        command_type,
-        dedupe_key,
-        command: Value::Object(payload),
-    }
-}
-
-#[cfg(test)]
-fn apply_runtime_activity_summary(
-    summary: &mut WorkflowRuntimeTreeSummary,
-    jobs_by_command: &BTreeMap<String, Vec<RuntimeJob>>,
-) {
-    for job in jobs_by_command.values().flatten() {
-        if let Some(outcome) = activity_result_envelope_from_job(job).and_then(|envelope| {
-            envelope
-                .get("outcome")
-                .and_then(Value::as_str)
-                .map(str::to_string)
-        }) {
-            summary
-                .activity_outcomes
-                .entry(outcome)
-                .and_modify(|count| *count += 1)
-                .or_insert(1);
-        } else {
-            summary.jobs_without_activity_envelope += 1;
-        }
-    }
-}
-
-fn prompt_packet_digest_from_events(events: &[RuntimeEvent]) -> Option<String> {
-    events.iter().rev().find_map(|event| {
-        if event.event_type != "RuntimePromptPrepared" {
-            return None;
-        }
-        event
-            .event
-            .get("prompt_packet_digest")
-            .and_then(Value::as_str)
-            .map(str::to_string)
-    })
-}
-
-fn runtime_job_has_in_flight_model_turn(job: &RuntimeJob, events: &[RuntimeEvent]) -> bool {
-    if job.status != RuntimeJobStatus::Running {
-        return false;
-    }
-    let Some(latest_turn_sequence) = events
-        .iter()
-        .filter(|event| event.event_type == "RuntimeTurnStarted")
-        .map(|event| event.sequence)
-        .max()
-    else {
-        return false;
-    };
-    !events.iter().any(|event| {
-        event.event_type == "ActivityResultReady" && event.sequence > latest_turn_sequence
-    })
-}
-
-fn last_runtime_observation_at(
-    job: &RuntimeJob,
-    events: &[RuntimeEvent],
-) -> Option<chrono::DateTime<chrono::Utc>> {
-    let latest_event_at = events.last().map(|event| event.created_at);
-    if job.status == RuntimeJobStatus::Running {
-        return Some(match latest_event_at {
-            Some(created_at) => created_at.max(job.updated_at),
-            None => job.updated_at,
-        });
-    }
-    latest_event_at
-}
-
-fn activity_result_envelope_from_job(job: &RuntimeJob) -> Option<Value> {
-    let artifacts = job.output.as_ref()?.get("artifacts")?.as_array()?;
-    artifacts.iter().rev().find_map(|artifact| {
-        if artifact.get("artifact_type").and_then(Value::as_str)
-            != Some(ACTIVITY_RESULT_ENVELOPE_ARTIFACT_TYPE)
-        {
-            return None;
-        }
-        let payload = artifact.get("artifact")?;
-        if payload.get("schema").and_then(Value::as_str) != Some(ACTIVITY_RESULT_ENVELOPE_SCHEMA) {
-            return None;
-        }
-        Some(payload.clone())
-    })
-}
-
-fn attach_workflow_children(
-    workflow_id: &str,
-    by_id: &BTreeMap<String, WorkflowRuntimeTreeNode>,
-    children_by_parent: &BTreeMap<String, Vec<String>>,
-    path: &mut BTreeSet<String>,
-) -> Option<WorkflowRuntimeTreeNode> {
-    if !path.insert(workflow_id.to_string()) {
-        return None;
-    }
-    let mut node = by_id.get(workflow_id)?.clone();
-    node.children = children_by_parent
-        .get(workflow_id)
-        .into_iter()
-        .flatten()
-        .filter_map(|child_id| attach_workflow_children(child_id, by_id, children_by_parent, path))
-        .collect();
-    path.remove(workflow_id);
-    Some(node)
-}
-
-#[cfg(test)]
-#[path = "misc_routes_runtime_tree_tests.rs"]
-mod runtime_tree_tests;
 
 pub(crate) async fn handle_rpc(
     State(state): State<Arc<AppState>>,
